hide full name, not only character name
With `HideNameInChat` enabled, hide the full display name (character
name + gamertag) instead of just the character name.

This allows not to have to install Multiplayer Tweaks on top of
IncognitoMode just for disabling gamertags, and also allows to keep
gamertags in the general case while still having IncognitoMode working
as intended.