progression: add item `IsActionBlocked` exception mechanism
In vanilla, it is possible to find Poison Arrows in Sunken Crypts in the
Swamp, even though they cannot be crafted until finding Obsidian in the
mountain.

This change adds an exception similar to the existing exceptions for
Entrails, allowing Poison Arrows to be used as soon as the right biome
is accessible, bypassing the recipe check.