progression: fix `Recipe`'s `IsActionBlocked` for one ingredient recipes
If `m_requireOnlyOneIngredient` is set, the recipe should be craftable
using any of the `Piece.Requirement` listed in `m_resources`. However,
this was not the case because `Recipe`'s `IsActionBlocked` was checking
all requirements even when `m_requireOnlyOneIngredient` was set.

We fix this with a special case to handle `m_requireOnlyOneIngredient`.